chore: 리뷰어님 코멘트 반영(20%)
- [ ] 4-8.리뷰어님 코멘트 반영
        - [x] 4-8-1.강의 자료 이미지 제거
        - [x] 4-8-2.HTTP Status NotFound로 변경
        - [ ] 4-8-3.`SectionGroup`생성
        - [ ] 4-8-4.`Line`엔티티의 필드로 `SectionGroup` 추가
        - [ ] 4-8-5.`Line`엔티티의 필드 중 `StationGroup` 제거
        - [ ] 4-8-6.`LineService`
            - [ ] 4-8-6-1.간결하게 하기
            - [ ] 4-8-6-2.
        - [ ] 4-8-7.구간을 순서대로 정렬하여 상행 종점부터 하행 종점까지 목록을 응답하기
        - [ ] 4-8-8.연관 관계 매핑 변경하기: `SectionGroup` 내 `Station` `@OneToOne` -> `@ManyToOne`
        - [x] 4-8-9.`Distance`: `double` -> `int`
        - [x] 4-8-10.`Distance.java`: 검증 메서드 중복 제거

Resolve: #309
See Also:

+++ src/main/java/nextstep/subway/section/domain/Section.java
@@ -41,7 +41,7 @@ public class Section extends BaseEntity {
 	protected Section() {
 	}
 
-	protected Section(Line line, Station upStation, Station downStation, double distance) {
+	protected Section(Line line, Station upStation, Station downStation, int distance) {
 		this(line, upStation, downStation, String.valueOf(distance));
 	}
 
@@ -55,7 +55,7 @@ public class Section extends BaseEntity {
 		this.line = line;
 		this.upStation = upStation;
 		this.downStation = downStation;
-		this.distance = Distance.generate(Double.parseDouble(distance));
+		this.distance = Distance.generate(Integer.parseInt(distance));
 	}
 
 	public static Section generate(Line line, Station upStation, Station downStation, String distance) {
@@ -64,7 +64,7 @@ public class Section extends BaseEntity {
 
 	private void validateDistanceParseDouble(String distance) {
 		try {
-			Double.parseDouble(distance);
+			Integer.parseInt(distance);
 		} catch (NumberFormatException e) {
 			throw new IllegalArgumentException("간격은 반드시 숫자로 입력되어야 합니다.");
 		}
@@ -116,7 +116,7 @@ public class Section extends BaseEntity {
 		return line;
 	}
 
-	public double distance() {
+	public int distance() {
 		return distance.distance();
 	}
